haystack-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
haystack-ai>=2.0.0
chroma-haystack>=0.17.0
ollama-haystack>=0.1.0
ollama==0.1.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
chromadb==0.4.22
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
spacy==3.7.2
reportlab==4.0.9
playwright==1.41.0
